made test. test failed. customerService is null. getting 22 internal calls

+++ src/main/java/com/example/SpringYard/repository/CustomerRepository.java
@@ -9,7 +9,7 @@ public interface CustomerRepository {
 
     public void updateCustomer(int id, Customer customer);
 
-    public Customer findCustomerById(int id);
+    public Customer findCustomerById(Long id);
 
     public List<Customer> getAllCustomers();
